requirements.txt
langchain
langchain-community
langchain-google-genai
wikipedia
duckduckgo-search
pydantic
python-dotenv